# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "GeoTiffMetadata": {
        "description": "Geospatial header fields read from a GeoTIFF (rasterio).\n\nEvery field is optional: a tile whose header cannot be parsed (or when the\ngeospatial stack is not installed) still lists, with these left null and a\nwarning on the item.",
        "properties": {
          "band_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Band Count"
          },
          "bounds": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bounds"
          },
          "crs": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crs"
          },
          "dtype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dtype"
          },
          "gsd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gsd"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          }
        },
        "title": "GeoTiffMetadata",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ImageryItem": {
        "description": "One tile in the Imagery Library.",
        "properties": {
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeoTiffMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "metadata_warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Warning"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "size_human": {
            "title": "Size Human",
            "type": "string"
          },
          "uploaded_at": {
            "format": "date-time",
            "title": "Uploaded At",
            "type": "string"
          }
        },
        "required": [
          "key",
          "filename",
          "size_bytes",
          "size_human",
          "uploaded_at"
        ],
        "title": "ImageryItem",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local API for Clay Geospatial Embeddings: run the Clay foundation model on satellite/aerial imagery tiles stored in Backblaze B2, write embedding tensors back to B2, and run k-NN similarity search over the archive. This contract documents the app's local API, not a hosted public endpoint.",
    "title": "Clay Geospatial Embeddings API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/library": {
      "get": {
        "operationId": "list_library_endpoint_library_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ImageryItem"
                  },
                  "title": "Response List Library Endpoint Library Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Library Endpoint",
        "tags": [
          "library"
        ]
      }
    },
    "/library/thumbnail": {
      "get": {
        "description": "PNG thumbnail for one imagery tile (referenced directly by <img>).",
        "operationId": "library_thumbnail_endpoint_library_thumbnail_get",
        "parameters": [
          {
            "in": "query",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 256,
              "title": "Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Library Thumbnail Endpoint",
        "tags": [
          "library"
        ]
      }
    },
...
  }
}